# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.SubmitReviewRequest:
    properties:
      content:
        minLength: 1
        type: string
      email:
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - content
    - email
    - name
    type: object
...
  models.ReviewResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/utils.PaginationMeta'
      success:
        type: boolean
    type: object
  utils.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Production-ready backend API for Kynettic platform with authentication,
    user management, and health monitoring
  title: Kynettic API
  version: "1.0"
paths:
...
  /reviews:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all reviews
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reviews retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ReviewResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List reviews
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Submit a review with name, email and content
      parameters:
      - description: Review details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SubmitReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Review submitted successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ReviewResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Submit a review
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"